Add Order, OrderItem interfaces
