+++ src/Contracts/Checkout/CartItemInterface.php
@@ -5,11 +5,9 @@ declare(strict_types=1);
 namespace Tipoff\Support\Contracts\Checkout;
 
 use Carbon\Carbon;
-use Tipoff\Support\Contracts\Models\BaseModelInterface;
-use Tipoff\Support\Contracts\Sellable\Sellable;
 use Tipoff\Support\Objects\DiscountableValue;
 
-interface CartItemInterface extends BaseModelInterface
+interface CartItemInterface extends BaseItemInterface
 {
     /**
      * Get cart containing this cart item
@@ -17,52 +15,37 @@ interface CartItemInterface extends BaseModelInterface
     public function getCart(): ?CartInterface;
 
     /**
-     * Get instance of Sellable associated w/this cart item
+     * Set item quantity
      */
-    public function getSellable(): Sellable;
-
-    /**
-     * Set Sellable provided opaque item id
-     */
-    public function getItemId(): string;
-
-    /**
-     * Get/Set item quantity
-     */
-    public function getQuantity(): int;
-
     public function setQuantity(int $qty): self;
 
     /**
-     * Get/Set user friendly description
+     * Set user friendly description
      */
-    public function getDescription(): string;
-
     public function setDescription(string $description): self;
 
     /**
-     * Get/set discountable amount for this cart item
+     * Set discountable amount for this cart item
      *
      * @param DiscountableValue|int $amount
      * @return $this
      */
     public function setAmount($amount): self;
 
-    public function getAmount(): DiscountableValue;
+    /**
+     * Set tax charged for this order item
+     */
+    public function setTax(int $tax): self;
 
     /**
-     * Get/set location for this cart item.  An exception is thrown if cart items
+     * Set location for this cart item.  An exception is thrown if cart items
      * have different locations.
      */
-    public function getLocationId(): ?int;
-
     public function setLocationId(?int $locationId): self;
 
     /**
-     * Get/set tax code for this item
+     * Set tax code for this item
      */
-    public function getTaxCode(): ?string;
-
     public function setTaxCode(?string $taxCode): self;
 
     /**
@@ -80,24 +63,4 @@ interface CartItemInterface extends BaseModelInterface
     public function getParentItem(): ?CartItemInterface;
 
     public function setParentItem(?CartItemInterface $parent): self;
-
-    /**
-     * Get meta data by optional key.  Dot notion in key paths is supported via Arr::get(...)
-     *
-     * @param string|null $key
-     * @param $default
-     * @return mixed
-     */
-    public function getMetaData(?string $key, $default);
-
-    /**
-     * Set an array item to a given value using "dot" notation.
-     *
-     * If no key is given to the method, the entire array will be replaced.
-     *
-     * @param string|null $key
-     * @param $value
-     * @return $this
-     */
-    public function setMetaData(?string $key, $value): self;
 }
